pyproject.toml
[project]
name = "accio"
version = "0.1.0"
description = """
Accio is a lightweight web crawler that, like the magical summoning spell 'Accio', fetches and processes web content based on natural language queries. 
It utilizes small language models (sLLM) to efficiently process and retrieve information from crawled web pages.

Key Features:
- Natural language based web content retrieval
- Context-window based content processing
- Memory-efficient in-memory caching
- Lightweight vector similarity search
- Optimized for single desktop usage

The project focuses on providing magical-like data retrieval while maintaining minimal resource usage through efficient content processing and smart caching strategies.
"""
authors = [
    {name = "yomapi", email = "yomapi@example.com"}  # Please update with actual email
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core crawling
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    # LLM
    "llama-cpp-python>=0.2.0",  # For running Mistral and other sLLMs
    # Caching and storage
    "cachetools>=5.3.0",
    # Utilities
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
    "pydantic>=2.5.0",
]

[project.packages]
find = { where = ["src"] }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
